feat(VoteRepository): 가게의 투표현황을 조회하는 기능 구현

+++ src/main/java/com/raota/domain/ramenShop/dto/VoteResultsDto.java
@@ -8,6 +8,6 @@ import lombok.Getter;
 public class VoteResultsDto {
     private Long menu_id;
     private String menu_name;
-    private int vote_count;
-    private double percentage;
+    private Long vote_count;
+    private Double percentage;
 }
